[FLINK-21844][runtime] Do not auto-configure maxParallelism in REACTIVE scheduling mode
In Reactive scheduling mode, automatically setting
the max parallelism of vertices if unconfigured
leads to poor user experience: when switching
between modes, savepoint compatibility would break;
max parallelism is not something that is naturally
thought in the context of schedulers.

This change consolidates configuring parallelism
settings for both reactive mode and the default mode,
as well as makes the JobVertex the source of truth
for what's been configured.

Signed-off-by: austin ce <austin.cawley@gmail.com>